New database relationship design

See merge request cs309/2024fall/3_roy_2!30

+++ Backend/demo/src/main/java/coms309/backEnd/demo/repository/PersonalTaskRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface PersonalTaskRepository extends JpaRepository<PersonalTask, Integer> {
+public interface PersonalTaskRepository extends JpaRepository<PersonalTask, Long> {
+    //public List<PersonalTask> findAllByUser(User user);
     public List<PersonalTask> findAllByUser(User user);
 }
